% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_consistency.R
\name{player_consistency}
\alias{player_consistency}
\title{NBA Player consistency estimates for a player during a season.}
\usage{
player_consistency(player, season, metrics = "basic",
  which_metrics_basic = c("MP", "GS", "FG", "FGA", "FG\%", "3P", "3PA",
  "3P\%", "FT", "FTA", "FT\%", "ORB", "DRB", "TRB", "AST", "STL", "BLK",
  "TOV", "PF", "PTS", "GmSc", "+/-"), which_metrics_advanced = c("MP",
  "GS", "TS\%", "eFG\%", "ORB\%", "DRB\%", "TRB\%", "AST\%", "STL\%",
  "BLK\%", "TOV\%", "USG\%", "ORtg", "DRtg", "GmSc"))
}
\arguments{
\item{player}{url link for player. For example, 't/townska01' for Karl Anthony Towns. Link can be found for desired player at 'https://www.basketball-reference.com/players/'}

\item{season}{A numeric year}

\item{metrics}{One of 'basic' or 'advanced'.}

\item{which_metrics_basic}{Any subset of ('MP', 'GS', 'FG', 'FGA', 'FG%', "3P","3PA", "3P%", "FT", "FTA", "FT%","ORB", "DRB", "TRB", "AST", "STL","BLK", "TOV", "PF", "PTS", "GmSc", "+/-").
Chosen statistics will be computed if metrics = 'basic'.}

\item{which_metrics_advanced}{Any subset of ('MP',"GS","TS%", "eFG%", 'ORB%', 'DRB%','TRB%', 'AST%',"STL%", "BLK%", "TOV%", "USG%", "ORtg","DRtg", 'GmSc'). Chosen statistics will be computed if metrics = 'advanced'}
}
\value{
A list with three element. The first element is a vector with the percent of games played and percent of games started out of total games played.
Second element is vector with average statistic per game. The third element is the variance of each statistic between games. Note if 'GS' is not included in which_metrics argument, the first element of list will not be computed.
}
\description{
Calculated consistency metrics for NBA player for a season
}
