% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_consistency_plot.R
\name{player_consistency_plot}
\alias{player_consistency_plot}
\title{NBA Player consistency plots for a player during a season.}
\usage{
player_consistency_plot(player, season, metrics = "basic",
  which_metrics_basic = c("MP", "FG", "FGA", "FG\%", "3P", "3PA", "3P\%",
  "FT", "FTA", "FT\%", "ORB", "DRB", "TRB", "AST", "STL", "BLK", "TOV",
  "PF", "PTS", "GmSc", "+/-"), which_metrics_advanced = c("MP", "TS\%",
  "eFG\%", "ORB\%", "DRB\%", "TRB\%", "AST\%", "STL\%", "BLK\%", "TOV\%",
  "USG\%", "ORtg", "DRtg", "GmSc"), plot_color = "navy",
  xlab = "Statistic", ylab = "", main = "", geom = "violin")
}
\arguments{
\item{player}{url link for player. For example, 't/townska01' for Karl Anthony Towns. Link can be found for desired player at 'https://www.basketball-reference.com/players/'}

\item{season}{A numeric year}

\item{metrics}{One of 'basic' or 'advanced'.}

\item{which_metrics_basic}{Any subset of ('MP', 'GS', 'FG', 'FGA', 'FG%', "3P","3PA", "3P%", "FT", "FTA", "FT%","ORB", "DRB", "TRB", "AST", "STL","BLK", "TOV", "PF", "PTS", "GmSc", "+/-")
Chosen statistics will be computed if metrics = 'basic'.}

\item{which_metrics_advanced}{Any subset of ('MP',"GS","TS%", "eFG%", 'ORB%', 'DRB%','TRB%', 'AST%',"STL%", "BLK%", "TOV%", "USG%", "ORtg","DRtg", 'GmSc')
Chosen statistics will be computed if metrics = 'advanced'}

\item{plot_color}{Color that works for ggplot object}

\item{xlab}{x-axis label}

\item{ylab}{y-axis lable}

\item{main}{title for plots}

\item{geom}{One of 'violin' or 'boxplot'.}
}
\value{
Four plots visualizing NBA players statistics.
}
\description{
Plots consistency metrics for NBA player for a season. See palyer_consistency() for details on statistics.
}
\examples{
player_consistency_plot('t/townska01', season = 2016, metrics = 'basic')

}
