% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{NBA Player per Game Stats for Season}
\usage{
player_stats(player, season, metrics = "basic")
}
\arguments{
\item{player}{url link for player. For example, 'j/jamesle01'. Link can be found for desired player at 'https://www.basketball-reference.com/players/'}

\item{season}{A numeric year}

\item{metrics}{One of 'basic' or 'advanced'. 'basic' will return basic statistics and 'advanced' will return advanced statistics}
}
\value{
An data frame containing player season game by game statistics from basketball-reference.com.
}
\description{
NBA Player per Game Stats for Season
}
\examples{
player_stats('c/curryst01', season = 2019, metrics = "advanced")
#returns Steph Curry's 2019 advanced statistics
}
